usr/requirements.txt
werkzeug==1.0.1
Flask==1.1.1
flask-restx==0.2
pymongo
pyzmq==19.0.0
flask-cors==3.0.8
pyopenssl